requirements.txt
igl==2.2.1
imageio==2.9.0
numpy==1.22.2
nvdiffrast==0.2.7
Pillow==9.2.0
plyfile==0.7.4
pytorch_lightning==1.4.1
PyYAML==6.0
scikit_image==0.18.2
scikit_learn==1.1.3
scipy==1.5.3
skimage==0.0
torch==1.7.1
trimesh==3.9.3
